+++ scripts/maintenance/computational-clusters/autoscaled_monitor/core.py
@@ -3,14 +3,10 @@
 import asyncio
 import datetime
 import json
-import re
-from collections import defaultdict
 from dataclasses import replace
 from pathlib import Path
-from typing import Any, Final
 
 import arrow
-import paramiko
 import parse
 import rich
 import typer
@@ -21,14 +17,14 @@ from rich.progress import track
 from rich.style import Style
 from rich.table import Column, Table
 
-from . import dask, db, ec2
+from . import dask, db, ec2, ssh, utils
+from .constants import SSH_USER_NAME, UNDEFINED_BYTESIZE
 from .models import (
     AppState,
     ComputationalCluster,
     ComputationalInstance,
     ComputationalTask,
     DaskTask,
-    DockerContainer,
     DynamicInstance,
     DynamicService,
     InstanceRole,
@@ -37,37 +33,11 @@ from .models import (
 )
 
 
-def _get_instance_name(instance) -> str:
-    for tag in instance.tags:
-        if tag["Key"] == "Name":
-            return tag.get("Value", "unknown")
-    return "unknown"
-
-
-def _get_last_heartbeat(instance) -> datetime.datetime | None:
-    for tag in instance.tags:
-        if tag["Key"] == "last_heartbeat":
-            return arrow.get(tag["Value"]).datetime
-    return None
-
-
-def _timedelta_formatting(
-    time_diff: datetime.timedelta, *, color_code: bool = False
-) -> str:
-    formatted_time_diff = f"{time_diff.days} day(s), " if time_diff.days > 0 else ""
-    formatted_time_diff += f"{time_diff.seconds // 3600:02}:{(time_diff.seconds // 60) % 60:02}:{time_diff.seconds % 60:02}"
-    if time_diff.days and color_code:
-        formatted_time_diff = f"[red]{formatted_time_diff}[/red]"
-    elif (time_diff.seconds > 5 * HOUR) and color_code:
-        formatted_time_diff = f"[orange]{formatted_time_diff}[/orange]"
-    return formatted_time_diff
-
-
 def _parse_computational(instance: Instance) -> ComputationalInstance | None:
-    name = _get_instance_name(instance)
+    name = utils.get_instance_name(instance)
     if result := state.computational_parser.search(name):
         assert isinstance(result, parse.Result)
-        last_heartbeat = _get_last_heartbeat(instance)
+        last_heartbeat = utils.get_last_heartbeat(instance)
         return ComputationalInstance(
             role=InstanceRole(result["role"]),
             user_id=result["user_id"],
@@ -96,7 +66,7 @@ def _create_graylog_permalinks(
 
 
 def _parse_dynamic(instance: Instance) -> DynamicInstance | None:
-    name = _get_instance_name(instance)
+    name = utils.get_instance_name(instance)
     if result := state.dynamic_parser.search(name):
         assert isinstance(result, parse.Result)
 
@@ -107,196 +77,6 @@ def _parse_dynamic(instance: Instance) -> DynamicInstance | None:
             disk_space=UNDEFINED_BYTESIZE,
         )
     return None
-
-
-def _ssh_and_get_dask_ip(
-    instance: Instance, username: str, private_key_path: Path
-) -> str:
-    # Establish SSH connection with key-based authentication
-    assert state.ssh_key_path
-    with _ssh_tunnel(
-        ssh_host=_BASTION_HOST,
-        username=username,
-        private_key_path=state.ssh_key_path,
-        remote_bind_host=instance.private_ip_address,
-        remote_bind_port=_DEFAULT_SSH_PORT,
-    ) as tunnel:
-        assert tunnel  # nosec
-        host, port = tunnel.local_bind_address
-        forward_url = f"tls://{host}:{port}"
-
-        with paramiko.SSHClient() as client:
-            client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-            try:
-                client.connect(
-                    forward_url,
-                    username=username,
-                    key_filename=f"{private_key_path}",
-                    timeout=5,
-                )
-                # Command to get disk space for /docker partition
-                dask_ip_command = "docker inspect -f '{{.NetworkSettings.Networks.dask_stack_default.IPAddress}}' $(docker ps --filter 'name=dask-sidecar|dask-scheduler' --format '{{.ID}}')"
-
-                # Run the command on the remote machine
-                _, stdout, _ = client.exec_command(dask_ip_command)
-                exit_status = stdout.channel.recv_exit_status()
-
-                if exit_status != 0:
-                    return "Not Found / Drained / Not Ready"
-
-                # Available disk space will be captured here
-                return stdout.read().decode("utf-8").strip()
-            except (
-                paramiko.AuthenticationException,
-                paramiko.SSHException,
-                TimeoutError,
-            ):
-                return "Not Ready"
-
-
-def _ssh_and_get_available_disk_space(
-    instance: Instance, username: str, private_key_path: Path
-) -> ByteSize:
-    assert state.ssh_key_path
-    with _ssh_tunnel(
-        ssh_host=_BASTION_HOST,
-        username=username,
-        private_key_path=state.ssh_key_path,
-        remote_bind_host=instance.private_ip_address,
-        remote_bind_port=_DEFAULT_SSH_PORT,
-    ) as tunnel:
-        assert tunnel  # nosec
-        host, port = tunnel.local_bind_address
-        forward_url = f"tls://{host}:{port}"
-        with paramiko.SSHClient() as client:
-            client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-            try:
-                client.connect(
-                    forward_url,
-                    username=username,
-                    key_filename=f"{private_key_path}",
-                    timeout=5,
-                )
-                # Command to get disk space for /docker partition
-                disk_space_command = (
-                    "df --block-size=1 /mnt/docker | awk 'NR==2{print $4}'"
-                )
-
-                # Run the command on the remote machine
-                _, stdout, stderr = client.exec_command(disk_space_command)
-                exit_status = stdout.channel.recv_exit_status()
-                error = stderr.read().decode()
-
-                if exit_status != 0:
-                    rich.print(error)
-                    raise typer.Abort(error)
-
-                # Available disk space will be captured here
-                available_space = stdout.read().decode("utf-8").strip()
-                return ByteSize(available_space)
-            except (
-                paramiko.AuthenticationException,
-                paramiko.SSHException,
-                TimeoutError,
-            ):
-                return ByteSize(0)
-
-
-def _ssh_and_list_running_dyn_services(
-    instance: Instance, username: str, private_key_path: Path
-) -> list[DynamicService]:
-    assert state.ssh_key_path
-    with _ssh_tunnel(
-        ssh_host=_BASTION_HOST,
-        username=username,
-        private_key_path=state.ssh_key_path,
-        remote_bind_host=instance.private_ip_address,
-        remote_bind_port=_DEFAULT_SSH_PORT,
-    ) as tunnel:
-        assert tunnel  # nosec
-        host, port = tunnel.local_bind_address
-        forward_url = f"tls://{host}:{port}"
-        with paramiko.SSHClient() as client:
-            client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-            try:
-                client.connect(
-                    forward_url,
-                    username=username,
-                    key_filename=f"{private_key_path}",
-                    timeout=5,
-                )
-                # Run the Docker command to list containers
-                _stdin, stdout, stderr = client.exec_command(
-                    'docker ps --format=\'{{.Names}}\t{{.CreatedAt}}\t{{.Label "io.simcore.runtime.user-id"}}\t{{.Label "io.simcore.runtime.project-id"}}\t{{.Label "io.simcore.name"}}\t{{.Label "io.simcore.version"}}\' --filter=name=dy-',
-                )
-                exit_status = stdout.channel.recv_exit_status()
-                error = stderr.read().decode()
-                if exit_status != 0:
-                    rich.print(error)
-                    raise typer.Abort(error)
-
-                output = stdout.read().decode("utf-8")
-                # Extract containers that follow the naming convention
-                running_service: dict[str, list[DockerContainer]] = defaultdict(list)
-                for container in output.splitlines():
-                    if match := re.match(DYN_SERVICES_NAMING_CONVENTION, container):
-                        named_container = DockerContainer(
-                            match["node_id"],
-                            int(match["user_id"]),
-                            match["project_id"],
-                            arrow.get(
-                                match["created_at"],
-                                "YYYY-MM-DD HH:mm:ss",
-                                tzinfo=datetime.timezone.utc,
-                            ).datetime,
-                            container,
-                            (
-                                json.loads(match["service_name"])["name"]
-                                if match["service_name"]
-                                else ""
-                            ),
-                            (
-                                json.loads(match["service_version"])["version"]
-                                if match["service_version"]
-                                else ""
-                            ),
-                        )
-                        running_service[match["node_id"]].append(named_container)
-
-                def _needs_manual_intervention(
-                    running_containers: list[DockerContainer],
-                ) -> bool:
-                    valid_prefixes = ["dy-sidecar_", "dy-proxy_", "dy-sidecar-"]
-                    for prefix in valid_prefixes:
-                        found = any(
-                            container.name.startswith(prefix)
-                            for container in running_containers
-                        )
-                        if not found:
-                            return True
-                    return False
-
-                return [
-                    DynamicService(
-                        node_id=node_id,
-                        user_id=containers[0].user_id,
-                        project_id=containers[0].project_id,
-                        created_at=containers[0].created_at,
-                        needs_manual_intervention=_needs_manual_intervention(
-                            containers
-                        ),
-                        containers=[c.name for c in containers],
-                        service_name=containers[0].service_name,
-                        service_version=containers[0].service_version,
-                    )
-                    for node_id, containers in running_service.items()
-                ]
-            except (
-                paramiko.AuthenticationException,
-                paramiko.SSHException,
-                TimeoutError,
-            ):
-                return []
 
 
 def _print_dynamic_instances(
@@ -339,7 +119,7 @@ def _print_dynamic_instances(
                     service.node_id,
                     service.service_name,
                     service.service_version,
-                    _timedelta_formatting(
+                    utils.timedelta_formatting(
                         time_now - service.created_at, color_code=True
                     ),
                     f"{'[red]' if service.needs_manual_intervention else ''}{service.needs_manual_intervention}{'[/red]' if service.needs_manual_intervention else ''}",
@@ -348,15 +128,15 @@ def _print_dynamic_instances(
         table.add_row(
             "\n".join(
                 [
-                    f"{_color_encode_with_state(instance.name, instance.ec2_instance)}",
+                    f"{utils.color_encode_with_state(instance.name, instance.ec2_instance)}",
                     f"ID: {instance.ec2_instance.instance_id}",
                     f"AMI: {instance.ec2_instance.image_id}",
                     f"AMI name: {instance.ec2_instance.image.name}",
                     f"Type: {instance.ec2_instance.instance_type}",
-                    f"Up: {_timedelta_formatting(time_now - instance.ec2_instance.launch_time, color_code=True)}",
+                    f"Up: {utils.timedelta_formatting(time_now - instance.ec2_instance.launch_time, color_code=True)}",
                     f"ExtIP: {instance.ec2_instance.public_ip_address}",
                     f"IntIP: {instance.ec2_instance.private_ip_address}",
-                    f"/mnt/docker(free): {_color_encode_with_threshold(instance.disk_space.human_readable(), instance.disk_space,  TypeAdapter(ByteSize).validate_python('15Gib'))}",
+                    f"/mnt/docker(free): {utils.color_encode_with_threshold(instance.disk_space.human_readable(), instance.disk_space,  TypeAdapter(ByteSize).validate_python('15Gib'))}",
                 ]
             ),
             service_table,
@@ -367,31 +147,6 @@ def _print_dynamic_instances(
             end_section=True,
         )
     rich.print(table, flush=True)
-
-
-def _get_worker_metrics(scheduler_info: dict[str, Any]) -> dict[str, Any]:
-    worker_metrics = {}
-    for worker_name, worker_data in scheduler_info.get("workers", {}).items():
-        worker_metrics[worker_name] = {
-            "resources": worker_data["resources"],
-            "tasks": worker_data["metrics"].get("task_counts", {}),
-        }
-    return worker_metrics
-
-
-def _color_encode_with_state(string: str, ec2_instance: Instance) -> str:
-    return (
-        f"[green]{string}[/green]"
-        if ec2_instance.state["Name"] == "running"
-        else f"[yellow]{string}[/yellow]"
-    )
-
-
-DANGER = "[red]{}[/red]"
-
-
-def _color_encode_with_threshold(string: str, value, threshold) -> str:
-    return string if value > threshold else DANGER.format(string)
 
 
 def _print_computational_clusters(
@@ -412,25 +167,25 @@ def _print_computational_clusters(
     for cluster in track(
         clusters, "Collecting information about computational clusters..."
     ):
-        cluster_worker_metrics = _get_worker_metrics(cluster.scheduler_info)
+        cluster_worker_metrics = dask.get_worker_metrics(cluster.scheduler_info)
         # first print primary machine info
         table.add_row(
             "\n".join(
                 [
-                    f"[bold]{_color_encode_with_state('Primary', cluster.primary.ec2_instance)}",
+                    f"[bold]{utils.color_encode_with_state('Primary', cluster.primary.ec2_instance)}",
                     f"Name: {cluster.primary.name}",
                     f"ID: {cluster.primary.ec2_instance.id}",
                     f"AMI: {cluster.primary.ec2_instance.image_id}",
                     f"AMI name: {cluster.primary.ec2_instance.image.name}",
                     f"Type: {cluster.primary.ec2_instance.instance_type}",
-                    f"Up: {_timedelta_formatting(time_now - cluster.primary.ec2_instance.launch_time, color_code=True)}",
+                    f"Up: {utils.timedelta_formatting(time_now - cluster.primary.ec2_instance.launch_time, color_code=True)}",
                     f"ExtIP: {cluster.primary.ec2_instance.public_ip_address}",
                     f"IntIP: {cluster.primary.ec2_instance.private_ip_address}",
                     f"DaskSchedulerIP: {cluster.primary.dask_ip}",
                     f"UserID: {cluster.primary.user_id}",
                     f"WalletID: {cluster.primary.wallet_id}",
-                    f"Heartbeat: {_timedelta_formatting(time_now - cluster.primary.last_heartbeat) if cluster.primary.last_heartbeat else 'n/a'}",
-                    f"/mnt/docker(free): {_color_encode_with_threshold(cluster.primary.disk_space.human_readable(), cluster.primary.disk_space,  TypeAdapter(ByteSize).validate_python('15Gib'))}",
+                    f"Heartbeat: {utils.timedelta_formatting(time_now - cluster.primary.last_heartbeat) if cluster.primary.last_heartbeat else 'n/a'}",
+                    f"/mnt/docker(free): {utils.color_encode_with_threshold(cluster.primary.disk_space.human_readable(), cluster.primary.disk_space,  TypeAdapter(ByteSize).validate_python('15Gib'))}",
                 ]
             ),
             "\n".join(
@@ -463,17 +218,17 @@ def _print_computational_clusters(
             table.add_row(
                 "\n".join(
                     [
-                        f"[italic]{_color_encode_with_state(f'Worker {index+1}', worker.ec2_instance)}[/italic]",
+                        f"[italic]{utils.color_encode_with_state(f'Worker {index+1}', worker.ec2_instance)}[/italic]",
                         f"Name: {worker.name}",
                         f"ID: {worker.ec2_instance.id}",
                         f"AMI: {worker.ec2_instance.image_id}",
                         f"AMI name: {worker.ec2_instance.image.name}",
                         f"Type: {worker.ec2_instance.instance_type}",
-                        f"Up: {_timedelta_formatting(time_now - worker.ec2_instance.launch_time, color_code=True)}",
+                        f"Up: {utils.timedelta_formatting(time_now - worker.ec2_instance.launch_time, color_code=True)}",
                         f"ExtIP: {worker.ec2_instance.public_ip_address}",
                         f"IntIP: {worker.ec2_instance.private_ip_address}",
                         f"DaskWorkerIP: {worker.dask_ip}",
-                        f"/mnt/docker(free): {_color_encode_with_threshold(worker.disk_space.human_readable(), worker.disk_space,  TypeAdapter(ByteSize).validate_python('15Gib'))}",
+                        f"/mnt/docker(free): {utils.color_encode_with_threshold(worker.disk_space.human_readable(), worker.disk_space,  TypeAdapter(ByteSize).validate_python('15Gib'))}",
                         "",
                     ]
                 ),
@@ -490,20 +245,22 @@ def _print_computational_clusters(
 
 
 async def _fetch_instance_details(
-    instance: DynamicInstance, ssh_key_path: Path
+    state: AppState, instance: DynamicInstance, ssh_key_path: Path
 ) -> tuple[list[DynamicService] | BaseException, ByteSize | BaseException]:
     # Run both SSH operations concurrently for this instance
     running_services, disk_space = await asyncio.gather(
         asyncio.get_event_loop().run_in_executor(
             None,
-            _ssh_and_list_running_dyn_services,
+            ssh.list_running_dyn_services,
+            state,
             instance.ec2_instance,
             SSH_USER_NAME,
             ssh_key_path,
         ),
         asyncio.get_event_loop().run_in_executor(
             None,
-            _ssh_and_get_available_disk_space,
+            ssh.get_available_disk_space,
+            state,
             instance.ec2_instance,
             SSH_USER_NAME,
             ssh_key_path,
@@ -514,13 +271,14 @@ async def _fetch_instance_details(
 
 
 async def _analyze_dynamic_instances_running_services_concurrently(
+    state: AppState,
     dynamic_instances: list[DynamicInstance],
     ssh_key_path: Path,
     user_id: int | None,
 ) -> list[DynamicInstance]:
     details = await asyncio.gather(
         *(
-            _fetch_instance_details(instance, ssh_key_path)
+            _fetch_instance_details(state, instance, ssh_key_path)
             for instance in dynamic_instances
         ),
         return_exceptions=True,
@@ -541,9 +299,6 @@ async def _analyze_dynamic_instances_running_services_concurrently(
     ]
 
 
-_SCHEDULER_PORT: Final[int] = 8786
-
-
 async def _analyze_computational_instances(
     state: AppState,
     computational_instances: list[ComputationalInstance],
@@ -556,7 +311,8 @@ async def _analyze_computational_instances(
             *(
                 asyncio.get_event_loop().run_in_executor(
                     None,
-                    _ssh_and_get_available_disk_space,
+                    ssh.get_available_disk_space,
+                    state,
                     instance.ec2_instance,
                     SSH_USER_NAME,
                     ssh_key_path,
@@ -570,7 +326,8 @@ async def _analyze_computational_instances(
             *(
                 asyncio.get_event_loop().run_in_executor(
                     None,
-                    _ssh_and_get_dask_ip,
+                    ssh.get_dask_ip,
+                    state,
                     instance.ec2_instance,
                     SSH_USER_NAME,
                     ssh_key_path,
@@ -653,6 +410,7 @@ async def _parse_computational_clusters(
 
 
 async def _parse_dynamic_instances(
+    state: AppState,
     instances: ServiceResourceInstancesCollection,
     ssh_key_path: Path | None,
     user_id: int | None,
@@ -666,7 +424,7 @@ async def _parse_dynamic_instances(
     if dynamic_instances and ssh_key_path:
         dynamic_instances = (
             await _analyze_dynamic_instances_running_services_concurrently(
-                dynamic_instances, ssh_key_path, user_id
+                state, dynamic_instances, ssh_key_path, user_id
             )
         )
     return dynamic_instances
@@ -679,7 +437,7 @@ async def summary(state: AppState, user_id: int | None, wallet_id: int | None) -
         state, user_id, wallet_id
     )
     dynamic_autoscaled_instances = await _parse_dynamic_instances(
-        dynamic_instances, state.ssh_key_path, user_id, wallet_id
+        state, dynamic_instances, state.ssh_key_path, user_id, wallet_id
     )
     _print_dynamic_instances(
         dynamic_autoscaled_instances,
@@ -692,7 +450,7 @@ async def summary(state: AppState, user_id: int | None, wallet_id: int | None) -
         state, user_id, wallet_id
     )
     computational_clusters = await _parse_computational_clusters(
-        computational_instances, state.ssh_key_path, user_id, wallet_id
+        state, computational_instances, state.ssh_key_path, user_id, wallet_id
     )
     _print_computational_clusters(
         computational_clusters,
@@ -745,11 +503,11 @@ async def _list_computational_clusters(
     state: AppState, user_id: int, wallet_id: int
 ) -> list[ComputationalCluster]:
     assert state.ec2_resource_clusters_keeper
-    computational_instances = await db.list_computational_instances_from_ec2(
+    computational_instances = await ec2.list_computational_instances_from_ec2(
         state, user_id, wallet_id
     )
     return await _parse_computational_clusters(
-        computational_instances, state.ssh_key_path, user_id, wallet_id
+        state, computational_instances, state.ssh_key_path, user_id, wallet_id
     )
 
 
@@ -853,7 +611,7 @@ async def trigger_cluster_termination(
         state, user_id, wallet_id
     )
     computational_clusters = await _parse_computational_clusters(
-        computational_instances, state.ssh_key_path, user_id, wallet_id
+        state, computational_instances, state.ssh_key_path, user_id, wallet_id
     )
     assert computational_clusters
     assert (
