+++ scripts/maintenance/computational-clusters/autoscaled_monitor/utils.py
@@ -1,0 +1,47 @@
+import datetime
+
+import arrow
+from mypy_boto3_ec2.service_resource import Instance
+
+from .constants import DANGER, HOUR
+
+
+def timedelta_formatting(
+    time_diff: datetime.timedelta, *, color_code: bool = False
+) -> str:
+    formatted_time_diff = f"{time_diff.days} day(s), " if time_diff.days > 0 else ""
+    formatted_time_diff += f"{time_diff.seconds // 3600:02}:{(time_diff.seconds // 60) % 60:02}:{time_diff.seconds % 60:02}"
+    if time_diff.days and color_code:
+        formatted_time_diff = f"[red]{formatted_time_diff}[/red]"
+    elif (time_diff.seconds > 5 * HOUR) and color_code:
+        formatted_time_diff = f"[orange]{formatted_time_diff}[/orange]"
+    return formatted_time_diff
+
+
+def get_instance_name(instance: Instance) -> str:
+    for tag in instance.tags:
+        assert "Key" in tag  # nosec
+        if tag["Key"] == "Name":
+            return tag.get("Value", "unknown")
+    return "unknown"
+
+
+def get_last_heartbeat(instance: Instance) -> datetime.datetime | None:
+    for tag in instance.tags:
+        assert "Key" in tag  # nosec
+        if tag["Key"] == "last_heartbeat":
+            assert "Value" in tag  # nosec
+            return arrow.get(tag["Value"]).datetime
+    return None
+
+
+def color_encode_with_state(string: str, ec2_instance: Instance) -> str:
+    return (
+        f"[green]{string}[/green]"
+        if ec2_instance.state["Name"] == "running"
+        else f"[yellow]{string}[/yellow]"
+    )
+
+
+def color_encode_with_threshold(string: str, value, threshold) -> str:
+    return string if value > threshold else DANGER.format(string)
